Add files via upload
To perform clustering on a customer dataset using machine learning, the following steps have been followed:
> Import the necessary libraries
>Load the dataset
>Data preprocessing
> Feature Selection
> Choosing the Clustering Algorithm
> Training the Model and assigning the clusters
> Evaluate and Visualize the Clusters
> Interpret the Results 
> Lastly, apply insights.
It's important to note that the choice of clustering algorithm and parameter tuning may vary depending on the dataset and problem at hand. Experimentation and iteration are key to finding the most suitable clustering approach for your customer dataset.